requirements.txt
datasets==1.18.4
discord.py==1.7.3
scikit_learn==1.0.2
sentence_transformers==2.2.0